requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
pyyaml>=6.0
openai>=1.30.0
redis>=5.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
apscheduler>=3.10.0
requests>=2.33.0
